# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * MyStoryBook API - AI-powered children's storybook generator
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Story {
  id: string;
  title: string;
  characterName: string;
  /** @nullable */
  characterName2?: string | null;
  relationship?: string;
  /** @nullable */
  relationship2?: string | null;
  theme?: string;
  age?: string;
  emotion?: string;
  /** @nullable */
  outfit?: string | null;
  pageCount?: number;
  /** @nullable */
  userPrompt?: string | null;
  /** @nullable */
  originalPhotoUrl?: string | null;
  /** @nullable */
  characterImageUrl?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** colour | colouring */
  style?: string;
  /** pending | generating | complete | error */
  status: string;
  generationProgress?: number;
  /** @nullable */
  generationStatusMessage?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}

export interface StoryInput {
  title: string;
  characterName: string;
  /** @nullable */
  characterName2?: string | null;
  relationship: string;
  /** @nullable */
  relationship2?: string | null;
  /** @nullable */
  petType?: string | null;
  /** @nullable */
  petType2?: string | null;
  theme: string;
  /** @nullable */
  customTheme?: string | null;
  age: string;
  emotion: string;
  /** @nullable */
  outfit?: string | null;
  /** @nullable */
  occasion?: string | null;
  pageCount: number;
  /** @nullable */
  userPrompt?: string | null;
  originalPhotoPath: string;
  /** @nullable */
  originalPhotoPath2?: string | null;
  /** colour | colouring */
  style?: string;
}

export interface StoryStatus {
  id: string;
  status: string;
  generationProgress: number;
  /** @nullable */
  generationStatusMessage?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
}

export interface StoryPage {
  id: string;
  storyId: string;
  pageNumber: number;
  /** @nullable */
  text?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  imagePrompt?: string | null;
}

export interface StoryReading {
  story: Story;
  pages: StoryPage[];
}

export interface LibraryStats {
  totalStories: number;
  completedStories: number;
  inProgressStories: number;
  recentStories?: Story[];
}

export interface GenerateResult {
  ok: boolean;
  /** @nullable */
  message?: string | null;
}

export interface DeleteResult {
  ok: boolean;
}