# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * MyStoryBook API - AI-powered children's storybook generator
 * OpenAPI spec version: 0.1.0
 */

export interface Story {
  id: string;
  title: string;
  characterName: string;
  /** @nullable */
  characterName2?: string | null;
  relationship?: string;
  /** @nullable */
  relationship2?: string | null;
  theme?: string;
  age?: string;
  emotion?: string;
  /** @nullable */
  outfit?: string | null;
  pageCount?: number;
  /** @nullable */
  userPrompt?: string | null;
  /** @nullable */
  originalPhotoUrl?: string | null;
  /** @nullable */
  characterImageUrl?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** colour | colouring */
  style?: string;
  /** pending | generating | complete | error */
  status: string;
  generationProgress?: number;
  /** @nullable */
  generationStatusMessage?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}