# lib/api-zod/src/generated/types/storyInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * MyStoryBook API - AI-powered children's storybook generator
 * OpenAPI spec version: 0.1.0
 */

export interface StoryInput {
  title: string;
  characterName: string;
  /** @nullable */
  characterName2?: string | null;
  relationship: string;
  /** @nullable */
  relationship2?: string | null;
  /** @nullable */
  petType?: string | null;
  /** @nullable */
  petType2?: string | null;
  theme: string;
  /** @nullable */
  customTheme?: string | null;
  age: string;
  emotion: string;
  /** @nullable */
  outfit?: string | null;
  /** @nullable */
  occasion?: string | null;
  pageCount: number;
  /** @nullable */
  userPrompt?: string | null;
  originalPhotoPath: string;
  /** @nullable */
  originalPhotoPath2?: string | null;
  /** colour | colouring */
  style?: string;
}